// frontend/app/src/generated/model/pathfinderItem.ts
/**
 * barley API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Usage } from './usage';
import { Price } from './price';
import { Publication } from './publication';
import { Level } from './level';
import { Material } from './material';
import { Traits } from './traits';


export interface PathfinderItem { 
    name?: any | null;
    type?: any | null;
    level?: Level;
    price?: Price;
    material?: Material;
    size?: any | null;
    usage?: Usage;
    traits?: Traits;
    publication?: Publication;
    quantity?: any | null;
}